cleanup(core): refactor process forking out of task orchestrator
