+++ packages/workspace/src/tasks-runner/task-orchestrator.ts
@@ -1,20 +1,18 @@
 import { Workspaces } from '@nrwl/tao/src/shared/workspace';
-import { ChildProcess, fork } from 'child_process';
-import * as dotenv from 'dotenv';
-import { readFileSync, writeFileSync } from 'fs';
+import { writeFileSync } from 'fs';
 import type { ProjectGraph } from '@nrwl/devkit';
-import { appRootPath } from '../utilities/app-root';
 import { output, TaskCacheStatus } from '../utilities/output';
 import { Cache, TaskWithCachedResult } from './cache';
 import { DefaultTasksRunnerOptions } from './default-tasks-runner';
 import { AffectedEventType, Task } from './tasks-runner';
-import { getOutputs, unparse } from './utils';
+import { getCommandArgsForTask, getExecutorForTask, getOutputs } from './utils';
 import { performance } from 'perf_hooks';
 import { TaskGraph } from './task-graph-creator';
 import { Hasher } from '../core/hasher/hasher';
+import { ForkedProcessTaskRunner } from './forked-process-task-runner';
+import { appRootPath } from '../utilities/app-root';
 
 export class TaskOrchestrator {
-  workspaceRoot = appRootPath;
   cache = new Cache(this.options);
   timings: { [target: string]: { start: number; end: number } } = {};
   completedTasks: {
@@ -27,7 +25,8 @@ export class TaskOrchestrator {
   waitingForTasks: Function[] = [];
   reverseTaskDeps: Record<string, string[]> = {};
 
-  private processes: ChildProcess[] = [];
+  private workspace = new Workspaces(appRootPath);
+  private forkedProcessTaskRunner = new ForkedProcessTaskRunner(this.options);
 
   constructor(
     private readonly hasher: Hasher,
@@ -36,9 +35,7 @@ export class TaskOrchestrator {
     private readonly taskGraph: TaskGraph,
     private readonly options: DefaultTasksRunnerOptions,
     private readonly hideCachedOutput: boolean
-  ) {
-    this.setupOnProcessExitListener();
-  }
+  ) {}
 
   async run() {
     this.calculateReverseDeps();
@@ -164,26 +161,15 @@ export class TaskOrchestrator {
         that.options.skipNxCache === false ||
         that.options.skipNxCache === undefined;
 
-      const cachedResult = await that.cache.get(task);
-      if (cachedResult && cachedResult.code === 0 && doNotSkipCache) {
-        that.applyCachedResult({ task, cachedResult });
-        await that.complete(task.id, 'cache');
-        return takeFromQueue();
+      if (doNotSkipCache && that.isCacheableTask(task)) {
+        const cachedResult = await that.cache.get(task);
+        if (cachedResult && cachedResult.code === 0) {
+          await that.applyCachedResult({ task, cachedResult });
+          return takeFromQueue();
+        }
       }
-
-      that.storeStartTime(task);
-      try {
-        const code = that.pipeOutputCapture(task)
-          ? await that.forkProcessPipeOutputCapture(task)
-          : await that.forkProcessDirectOutputCapture(task);
-
-        that.storeEndTime(task);
-        await that.complete(task.id, code === 0 ? 'success' : 'failure');
-        return takeFromQueue();
-      } catch {
-        await that.complete(task.id, 'failure');
-        return takeFromQueue();
-      }
+      await that.runTaskInForkedProcess(task);
+      return takeFromQueue();
     }
 
     const wait = [];
@@ -226,7 +212,46 @@ export class TaskOrchestrator {
     });
   }
 
-  private applyCachedResult(t: TaskWithCachedResult) {
+  private async runTaskInForkedProcess(task: Task) {
+    try {
+      this.storeStartTime(task);
+      this.options.lifeCycle.startTask(task);
+      const taskOutputs = getOutputs(this.projectGraph.nodes, task);
+      const outputPath = this.cache.temporaryOutputPath(task);
+      const forwardOutput = this.shouldForwardOutput(task);
+      this.cache.removeRecordedOutputsHashes(taskOutputs);
+      const pipeOutput = this.pipeOutputCapture(task);
+      const { code, terminalOutput } = pipeOutput
+        ? await this.forkedProcessTaskRunner.forkProcessPipeOutputCapture(
+            task,
+            {
+              forwardOutput,
+            }
+          )
+        : await this.forkedProcessTaskRunner.forkProcessDirectOutputCapture(
+            task,
+            {
+              temporaryOutputPath: outputPath,
+              forwardOutput,
+            }
+          );
+
+      this.storeEndTime(task);
+      if (this.shouldCacheTaskResult(task, code)) {
+        await this.cache.put(task, terminalOutput, taskOutputs, code);
+        this.cache.recordOutputsHash(taskOutputs, task.hash);
+        if (pipeOutput) {
+          writeFileSync(outputPath, terminalOutput);
+        }
+      }
+      this.options.lifeCycle.endTask(task, code);
+      await this.complete(task.id, code === 0 ? 'success' : 'failure');
+    } catch (e) {
+      await this.complete(task.id, 'failure');
+    }
+  }
+
+  private async applyCachedResult(t: TaskWithCachedResult) {
     this.storeStartTime(t.task);
     this.options.lifeCycle.startTask(t.task);
     const outputs = getOutputs(this.projectGraph.nodes, t.task);
@@ -242,7 +267,7 @@ export class TaskOrchestrator {
         this.initiatingProject === t.task.target.project) &&
       !this.hideCachedOutput
     ) {
-      const args = this.getCommandArgs(t.task);
+      const args = getCommandArgsForTask(t.task);
       output.logCommand(
         `nx ${args.join(' ')}`,
         shouldCopyOutputsFromCache
@@ -253,6 +278,7 @@ export class TaskOrchestrator {
     }
     this.options.lifeCycle.endTask(t.task, t.cachedResult.code);
     this.storeEndTime(t.task);
+    await this.complete(t.task.id, 'cache');
   }
 
   private storeStartTime(t: Task) {
@@ -268,17 +294,9 @@ export class TaskOrchestrator {
     ].end = new Date().getTime();
   }
 
-  private pipeOutputCapture(task: Task) {
-    try {
-      return this.readExecutor(task).schema.outputCapture === 'pipe';
-    } catch (e) {
-      return false;
-    }
-  }
-
   private customHasher(task: Task) {
     try {
-      const f = this.readExecutor(task).hasherFactory;
+      const f = getExecutorForTask(task, this.workspace).hasherFactory;
       return f ? f() : null;
     } catch (e) {
       console.error(e);
@@ -286,283 +304,42 @@ export class TaskOrchestrator {
     }
   }
 
-  private readExecutor(task: Task) {
-    const p = this.projectGraph.nodes[task.target.project];
-    const b = p.data.targets[task.target.target].executor;
-    const [nodeModule, executor] = b.split(':');
-
-    const w = new Workspaces(this.workspaceRoot);
-    return w.readExecutor(nodeModule, executor);
-  }
-
-  private forkProcessPipeOutputCapture(task: Task) {
-    const taskOutputs = getOutputs(this.projectGraph.nodes, task);
-    const outputPath = this.cache.temporaryOutputPath(task);
-    return new Promise((res, rej) => {
-      try {
-        this.options.lifeCycle.startTask(task);
-        const forwardOutput = this.shouldForwardOutput(outputPath, task);
-        const env = this.envForForkedProcess(
-          task,
-          undefined,
-          forwardOutput,
-          process.env.FORCE_COLOR === undefined
-            ? 'true'
-            : process.env.FORCE_COLOR
-        );
-        const args = this.getCommandArgs(task);
-        const commandLine = `nx ${args.join(' ')}`;
-
-        if (forwardOutput) {
-          output.logCommand(commandLine);
-        }
-        this.cache.removeRecordedOutputsHashes(taskOutputs);
-        const p = fork(this.getCommand(), args, {
-          stdio: ['inherit', 'pipe', 'pipe', 'ipc'],
-          env,
-        });
-        this.processes.push(p);
-        let out = [];
-        let outWithErr = [];
-        p.stdout.on('data', (chunk) => {
-          if (forwardOutput) {
-            process.stdout.write(chunk);
-          }
-          out.push(chunk.toString());
-          outWithErr.push(chunk.toString());
-        });
-        p.stderr.on('data', (chunk) => {
-          if (forwardOutput) {
-            process.stderr.write(chunk);
-          }
-          outWithErr.push(chunk.toString());
-        });
-
-        p.on('exit', (code, signal) => {
-          if (code === null) code = this.signalToCode(signal);
-          // we didn't print any output as we were running the command
-          // print all the collected output|
-          if (!forwardOutput) {
-            output.logCommand(commandLine);
-            process.stdout.write(outWithErr.join(''));
-          }
-          if (outputPath) {
-            const terminalOutput = outWithErr.join('');
-            writeFileSync(outputPath, terminalOutput);
-            if (this.shouldCacheTask(outputPath, code)) {
-              this.cache
-                .put(task, terminalOutput, taskOutputs, code)
-                .then(() => {
-                  this.cache.recordOutputsHash(taskOutputs, task.hash);
-                  this.options.lifeCycle.endTask(task, code);
-                  res(code);
-                })
-                .catch((e) => {
-                  rej(e);
-                });
-            } else {
-              this.cache.recordOutputsHash(taskOutputs, task.hash);
-              this.options.lifeCycle.endTask(task, code);
-              res(code);
-            }
-          } else {
-            this.cache.recordOutputsHash(taskOutputs, task.hash);
-            this.options.lifeCycle.endTask(task, code);
-            res(code);
-          }
-        });
-      } catch (e) {
-        console.error(e);
-        rej(e);
-      }
-    });
-  }
-
-  private forkProcessDirectOutputCapture(task: Task) {
-    const taskOutputs = getOutputs(this.projectGraph.nodes, task);
-    const outputPath = this.cache.temporaryOutputPath(task);
-    return new Promise((res, rej) => {
-      try {
-        this.options.lifeCycle.startTask(task);
-        const forwardOutput = this.shouldForwardOutput(outputPath, task);
-        const env = this.envForForkedProcess(
-          task,
-          outputPath,
-          forwardOutput,
-          undefined
-        );
-        const args = this.getCommandArgs(task);
-        const commandLine = `nx ${args.join(' ')}`;
-
-        if (forwardOutput) {
-          output.logCommand(commandLine);
-        }
-        this.cache.removeRecordedOutputsHashes(taskOutputs);
-        const p = fork(this.getCommand(), args, {
-          stdio: ['inherit', 'inherit', 'inherit', 'ipc'],
-          env,
-        });
-        this.processes.push(p);
-        p.on('exit', (code, signal) => {
-          if (code === null) code = this.signalToCode(signal);
-          // we didn't print any output as we were running the command
-          // print all the collected output
-          if (!forwardOutput) {
-            output.logCommand(commandLine);
-            const terminalOutput = this.readTerminalOutput(outputPath);
-            if (terminalOutput) {
-              process.stdout.write(terminalOutput);
-            } else {
-              console.error(
-                `Nx could not find process's output. Run the command without --parallel.`
-              );
-            }
-          }
-          // we don't have to worry about this statement. code === 0 guarantees the file is there.
-          if (this.shouldCacheTask(outputPath, code)) {
-            this.cache
-              .put(task, this.readTerminalOutput(outputPath), taskOutputs, code)
-              .then(() => {
-                this.cache.recordOutputsHash(taskOutputs, task.hash);
-                this.options.lifeCycle.endTask(task, code);
-                res(code);
-              })
-              .catch((e) => {
-                rej(e);
-              });
-          } else {
-            this.cache.recordOutputsHash(taskOutputs, task.hash);
-            this.options.lifeCycle.endTask(task, code);
-            res(code);
-          }
-        });
-      } catch (e) {
-        console.error(e);
-        rej(e);
-      }
-    });
-  }
-
-  private readTerminalOutput(outputPath: string) {
+  private pipeOutputCapture(task: Task) {
     try {
-      return readFileSync(outputPath).toString();
+      return (
+        getExecutorForTask(task, this.workspace).schema.outputCapture === 'pipe'
+      );
     } catch (e) {
-      return null;
+      return false;
     }
   }
 
-  private shouldCacheTask(outputPath: string | null, code: number) {
+  private shouldCacheTaskResult(task: Task, code: number) {
     // TODO: vsavkin make caching failures the default in Nx 12.1
-    if (process.env.NX_CACHE_FAILURES == 'true') {
-      return outputPath;
-    } else {
-      return outputPath && code === 0;
-    }
+    return (
+      this.isCacheableTask(task) &&
+      (process.env.NX_CACHE_FAILURES == 'true' ? true : code === 0)
+    );
   }
 
-  private envForForkedProcess(
-    task: Task,
-    outputPath: string,
-    forwardOutput: boolean,
-    forceColor: string
-  ) {
-    const envsFromFiles = {
-      ...parseEnv('.env'),
-      ...parseEnv('.local.env'),
-      ...parseEnv('.env.local'),
-      ...parseEnv(`${task.projectRoot}/.env`),
-      ...parseEnv(`${task.projectRoot}/.local.env`),
-      ...parseEnv(`${task.projectRoot}/.env.local`),
-    };
-
-    const env: NodeJS.ProcessEnv = {
-      ...envsFromFiles,
-      FORCE_COLOR: forceColor,
-      ...process.env,
-      NX_TASK_HASH: task.hash,
-      NX_INVOKED_BY_RUNNER: 'true',
-      NX_WORKSPACE_ROOT: this.workspaceRoot,
-    };
-
-    if (outputPath) {
-      env.NX_TERMINAL_OUTPUT_PATH = outputPath;
-      if (this.options.captureStderr) {
-        env.NX_TERMINAL_CAPTURE_STDERR = 'true';
-      }
-      // TODO: remove this once we have a reasonable way to configure it
-      if (task.target.target === 'test') {
-        env.NX_TERMINAL_CAPTURE_STDERR = 'true';
-      }
-      if (forwardOutput) {
-        env.NX_FORWARD_OUTPUT = 'true';
-      }
-    }
-    return env;
-  }
-
-  private shouldForwardOutput(outputPath: string | undefined, task: Task) {
-    if (!outputPath) return true;
+  private shouldForwardOutput(task: Task) {
+    if (!this.isCacheableTask(task)) return true;
     if (!this.options.parallel) return true;
     if (task.target.project === this.initiatingProject) return true;
     return false;
   }
 
-  private getCommand() {
-    const cli = require.resolve(`@nrwl/cli/lib/run-cli.js`, {
-      paths: [this.workspaceRoot],
-    });
-    return `${cli}`;
+  private isCacheableTask(task: Task) {
+    const cacheable =
+      this.options.cacheableOperations || this.options.cacheableTargets;
+    return (
+      cacheable &&
+      cacheable.indexOf(task.target.target) > -1 &&
+      !this.longRunningTask(task)
+    );
   }
 
-  private getCommandArgs(task: Task) {
-    const args: string[] = unparse(task.overrides || {});
-
-    const config = task.target.configuration
-      ? `:${task.target.configuration}`
-      : '';
-
-    return [
-      'run',
-      `${task.target.project}:${task.target.target}${config}`,
-      ...args,
-    ];
+  private longRunningTask(task: Task) {
+    return !!task.overrides['watch'];
   }
-
-  private setupOnProcessExitListener() {
-    process.on('SIGINT', () => {
-      this.processes.forEach((p) => {
-        p.kill('SIGTERM');
-      });
-      // we exit here because we don't need to write anything to cache.
-      process.exit();
-    });
-    process.on('SIGTERM', () => {
-      this.processes.forEach((p) => {
-        p.kill('SIGTERM');
-      });
-      // no exit here because we expect child processes to terminate which
-      // will store results to the cache and will terminate this process
-    });
-    process.on('SIGHUP', () => {
-      this.processes.forEach((p) => {
-        p.kill('SIGTERM');
-      });
-      // no exit here because we expect child processes to terminate which
-      // will store results to the cache and will terminate this process
-    });
-  }
-
-  private signalToCode(signal: string) {
-    if (signal === 'SIGHUP') return 128 + 1;
-    if (signal === 'SIGINT') return 128 + 2;
-    if (signal === 'SIGTERM') return 128 + 15;
-    return 128;
-  }
-}
-
-function parseEnv(path: string) {
-  try {
-    const envContents = readFileSync(path);
-    return dotenv.parse(envContents);
-  } catch (e) {}
 }
