+++ packages/workspace/src/tasks-runner/utils.ts
@@ -7,6 +7,7 @@ import {
 import { Task } from './tasks-runner';
 import { flatten } from 'flat';
 import { output } from '../utilities/output';
+import { Workspaces } from '@nrwl/tao/src/shared/workspace';
 
 const commonCommands = ['build', 'test', 'lint', 'e2e', 'deploy'];
 
@@ -198,4 +199,35 @@ function interpolateOutputs(template: string, data: any): string {
 
     return value;
   });
+}
+
+export function getExecutorForTask(task: Task, workspace: Workspaces) {
+  const project = workspace.readWorkspaceConfiguration().projects[
+    task.target.project
+  ];
+  const executor = project.targets[task.target.target].executor;
+  const [nodeModule, executorName] = executor.split(':');
+
+  return workspace.readExecutor(nodeModule, executorName);
+}
+
+export function getCliPath(workspaceRoot: string) {
+  const cli = require.resolve(`@nrwl/cli/lib/run-cli.js`, {
+    paths: [workspaceRoot],
+  });
+  return `${cli}`;
+}
+
+export function getCommandArgsForTask(task: Task) {
+  const args: string[] = unparse(task.overrides || {});
+
+  const config = task.target.configuration
+    ? `:${task.target.configuration}`
+    : '';
+
+  return [
+    'run',
+    `${task.target.project}:${task.target.target}${config}`,
+    ...args,
+  ];
 }
